+++ docs/changelog.md
@@ -17,6 +17,7 @@ Each release rewinds and fast-forwards your mixtape experience—just like a rea
 - Modal for statistics on the entire collection
 - Improved details in editor for compilation albums
 - Enlarged album cover in album details
+- Hierarchical hints in search results
 
 ### 🔧 Changed